+++ src/main/java/wooteco/chess/ConsoleApplication.java
@@ -1,6 +1,6 @@
 package wooteco.chess;
 
-import wooteco.chess.contoller.ConsoleController;
+import wooteco.chess.controller.ConsoleController;
 
 public class ConsoleApplication {
 	public static void main(String[] args) {
