+++ src/main/java/wooteco/chess/controller/SpringController.java
@@ -1,4 +1,4 @@
-package wooteco.chess.contoller;
+package wooteco.chess.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
