+++ src/main/java/wooteco/chess/dao/RoomDao.java
@@ -1,4 +1,4 @@
-package wooteco.chess.domain.dao;
+package wooteco.chess.dao;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
